code-of-conduct: Fix the ambiguity about collecting email addresses
The current code of conduct has an ambiguity in the it considers publishing
private information such as email addresses unacceptable behaviour.  Since
the Linux kernel collects and publishes email addresses as part of the patch
process, add an exception clause for email addresses ordinarily collected by
the project to correct this ambiguity.

Fixes: 8a104f8b5867c682 ("Code of Conduct: Let's revamp it.")
Acked-by: Geert Uytterhoeven <geert@linux-m68k.org>
Acked-by: Shuah Khan <shuah@kernel.org>
Acked-by: Guenter Roeck <linux@roeck-us.net>
Reviewed-by: Alan Cox <alan@llwyncelyn.cymru>
Reviewed-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>
Acked-by: "Eric W. Biederman" <ebiederm@xmission.com>
Acked-by: Kees Cook <keescook@chromium.org>
Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>