code-of-conduct: Add back the TAB as the central reporting point
Add a Reporting section where the Enforcement section used to be.  The
intention is still to debate what should go here, but in the meantime, this
gives us back the central reporting point we had in the old code of conflict.

Reviewed-by: Konrad Rzeszutek Wilk <konrad.wilk@oracle.com>
Reviewed-by: Mauro Carvalho Chehab <mchehab@kernel.org>
Acked-by: Geert Uytterhoeven <geert@linux-m68k.org>
Acked-by: Shuah Khan <shuah@kernel.org>
Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>

---

v2: Added this patch to allay concerns we were stripping the reporting
    mechanism entirely.
v3: Add (where required by law) as a qualifier to the confidentiality
    requirement